+++ src/dash/data.py
@@ -2,9 +2,10 @@
 from sqlalchemy import create_engine, MetaData, Table, select, join
 import pandas as pd
 import re
+import numpy as np
 
 # SQLite path
-db_path = 'sqlite:///../../data/SQLite Database/20200527/Covid-19 Study DB.sqlite'
+db_path = 'sqlite:///../../data/SQLite Database/20200528/Covid-19 Study DB.sqlite'
 
 omics_id_dict = {
         "proteomics":1,
@@ -185,3 +186,119 @@ def get_combined_data(df_dict, quant_range_dict):
     quant_range_dict['combined'] = combined_quant_range
 
     return df_dict, quant_range_dict
+
+def get_p_values():
+
+    # Create an engine that connects to the Covid-19 Study DB.sqlite file: engine
+    engine = create_engine(db_path)
+
+    # Establish connection
+    connection = engine.connect()
+
+    query = "SELECT * from biomolecules WHERE KEEP=1"
+    # get biomolecule names
+    biomolecules_df = pd.read_sql_query(query, connection)
+
+    # build biomolecule name dict and drop list
+    biomolecule_name_dict = {}
+    for index, row in biomolecules_df.iterrows():
+        biomolecule_id = str(row['biomolecule_id'])
+        standardized_name = row['standardized_name']
+        biomolecule_name_dict[biomolecule_id] = standardized_name
+
+    query = "SELECT * from pvalues"
+    # get biomolecule names
+    pvalues_df = pd.read_sql_query(query, connection)
+
+    return pvalues_df
+
+def get_volcano_data(pvalues_df, df_dict, quant_value_range,
+    global_names_dict, comparison_column='COVID'):
+
+    group_1_quant_value_dict = {}
+    comparison_column = 'COVID'
+
+    group_1 = "1"
+    group_2 = "0"
+
+    combined_df = df_dict['combined']
+    quant_value_columns = combined_df.columns[:quant_value_range]
+
+    group_1_quant_value_dict = {}
+
+    for sample_id, row in combined_df[combined_df[comparison_column]==group_1].iterrows():
+
+        for biomolecule_id in quant_value_columns:
+            quant_value = row[biomolecule_id]
+
+            if not biomolecule_id in group_1_quant_value_dict:
+                group_1_quant_value_dict[biomolecule_id] = [quant_value]
+
+            else:
+                group_1_quant_value_dict[biomolecule_id].append(quant_value)
+
+    group_2_quant_value_dict = {}
+
+    for sample_id, row in combined_df[combined_df[comparison_column]==group_2].iterrows():
+
+        for biomolecule_id in quant_value_columns:
+            quant_value = row[biomolecule_id]
+
+            if not biomolecule_id in group_2_quant_value_dict:
+                group_2_quant_value_dict[biomolecule_id] = [quant_value]
+
+            else:
+                group_2_quant_value_dict[biomolecule_id].append(quant_value)
+
+
+    FC_dict = {}
+    for biomolecule_id in quant_value_columns:
+
+        group_1_quant_values = group_1_quant_value_dict[biomolecule_id]
+        group_2_quant_values = group_2_quant_value_dict[biomolecule_id]
+
+        # in log2 space; subtract
+        FC = np.mean(group_1_quant_values) - np.mean(group_2_quant_values)
+
+        FC_dict[biomolecule_id] = FC
+
+    FC_list = []
+    ome_list = []
+    standardized_name_list = []
+    for index, row in pvalues_df.iterrows():
+        biomolecule_id = str(row['biomolecule_id'])
+
+        ## NOTE: should create biomolecule_ome_dict
+        if biomolecule_id in df_dict['proteomics'].columns:
+            ome_list.append("proteomics")
+        elif biomolecule_id in df_dict['lipidomics'].columns:
+            ome_list.append("lipidomics")
+        elif biomolecule_id in df_dict['metabolomics'].columns:
+            ome_list.append("metabolomics")
+        else:
+            #print("Biomolecule {} not mapped to ome!".format(biomolecule_id)) # don't currenly have targeted metabolomics included
+            ome_list.append(np.nan)
+            #break
+
+        try:
+            FC = FC_dict[biomolecule_id]
+        except:
+            # may have been a dropped biomolecule or from a different ome
+            FC = np.nan
+
+        FC_list.append(FC)
+
+        try:
+            standardized_name = global_names_dict['combined'][biomolecule_id]
+        except:
+            # may have been a dropped biomolecule or from a different ome
+            standardized_name = np.nan
+
+        standardized_name_list.append(standardized_name)
+
+    pvalues_df['log2_FC'] = FC_list
+    pvalues_df['ome_type'] = ome_list
+    pvalues_df['standardized_name'] = standardized_name_list
+    pvalues_df['neg_log10_p_value'] = pvalues_df['p_value'].apply(np.log10).apply(np.negative)
+
+    return pvalues_df