+++ src/dash/index.py
@@ -28,4 +28,6 @@ def display_page(pathname):
         return '404'
 
 if __name__ == '__main__':
-    app.run_server(debug=True)
+    app.run_server(
+        debug=True,
+        host='0.0.0.0')